requirements.txt
appdirs==1.4.3
attrs==19.1.0
cached-property==1.5.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
defusedxml==0.6.0
Flask==1.0.3
Flask-Failsafe==0.2
icalendar==4.0.3
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.3.3
MarkupSafe==1.1.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
requests-toolbelt==0.9.1
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.4
